+++ frontend/src/components/LanguageSwitcher.tsx
@@ -20,12 +20,12 @@ export default function LanguageSwitcher() {
   };
 
   return (
-    <div>
+    <span>
       <Tooltip title={t('header.change-language').toString()}>
-        <Button aria-controls="language-menu" aria-haspopup="true" onClick={handleClick}
-                style={{paddingLeft: 0, paddingRight: 0}}>
-          <TranslateIcon />
-          <ExpandMoreIcon />
+        <Button size="medium" aria-controls="language-menu" aria-haspopup="true" onClick={handleClick}
+                style={{minWidth: 0}}>
+          <TranslateIcon fontSize="small" />
+          <ExpandMoreIcon fontSize="small" />
         </Button>
       </Tooltip>
       <Menu id="language-menu" anchorEl={anchorEl} keepMounted open={open} onClose={handleClose}
@@ -40,6 +40,6 @@ export default function LanguageSwitcher() {
           {t('header.zh-CN')}
         </MenuItem>
       </Menu>
-    </div>
+    </span>
   );
 }