+++ frontend/src/pages/Home.tsx
@@ -9,7 +9,6 @@ import { GlobalDataContext } from '../GlobalDataContext';
 import { ResponseMapConnections, ResponseMessage, ResponseRouteFind, Waypoint } from '../response';
 import SystemInput from '../components/SystemInput';
 import RouteList from '../components/RouteList';
-import NavModal from '../components/NavModal';
 import Map from '../components/Map';
 
 const styles = (theme: Theme) => createStyles({
@@ -51,8 +50,6 @@ class Home extends React.Component<Props, HomeState> {
 
     return (
       <div className='grid-spacing-2-wrapper'>
-
-        <NavModal connectionChanged={this.connectionChanged} />
 
         <Grid container spacing={2} className='card'>
           <Grid item xs={12}>
@@ -136,14 +133,6 @@ class Home extends React.Component<Props, HomeState> {
     this.routeSet = this.routeSet.bind(this);
   }
 
-  componentDidMount() {
-    loadConnections(this);
-  }
-
-  connectionChanged = () => {
-    loadConnections(this)
-  };
-
   startChanged = (value: string) => {
     this.setState({routeFrom: value});
     this.setState({startSystemInput: value});
@@ -216,17 +205,3 @@ class Home extends React.Component<Props, HomeState> {
 }
 
 export default withTranslation()(withStyles(styles)(Home));
-
-/**
- * Load Ansiblex and temporary connections.
- */
-const loadConnections = (app: Home) => {
-  axios.get<ResponseMapConnections>(`${app.context.domain}/api/route/map-connections`).then(r => {
-    if (r.data.code) { // error
-      console.log(app.t(`responseCode.${r.data.code}`));
-    }
-    app.setState({mapConnections: r.data});
-  }).catch(() => {
-    app.setState({mapConnections: {ansiblexes: [], temporary: [], code: null}});
-  });
-};